+++ bt/app/ui/btwidgets.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "grey.h"
+#include "model.h"
 #include "../browser.h"
 
 /**
@@ -10,5 +11,10 @@ namespace bt::ui {
 
     void btw_icon(grey::app& app,
         const profile_selection& selection,
+        float padding, float icon_size, bool is_active);
+
+    void btw_icon(grey::app& app,
+        const browser& b, const browser_profile& p,
+        icon_overlay_mode icon_mode,
         float padding, float icon_size, bool is_active);
 }